refactor: rename schema files to *.sql.ts
`src/schema/projects.ts` and `src/projects/` were indistinguishable in
editor file search. The suffix disambiguates them, and lets drizzle.config
find table definitions by glob (`./src/**/*.sql.ts`) instead of by folder —
the prerequisite for colocating each table with the service that owns it.

Verified with `drizzle-kit generate`: the glob resolves to the same schema
as the previous `./src/schema/index.ts` entry point ("No schema changes,
nothing to migrate").

Part of #475.
