+++ packages/core/src/schema/daily-featured-projects.sql.ts
