+++ packages/core/src/schema/index.ts
@@ -1,12 +1,12 @@
-export * from "./bundles";
-export * from "./daily-featured-projects";
-export * from "./hall-of-fame";
-export * from "./packages";
-export * from "./project-trends";
-export * from "./project-trends-view";
-export * from "./projects";
-export * from "./repo-trends";
-export * from "./repo-trends-view";
-export * from "./repos";
-export * from "./snapshots";
-export * from "./tags";
+export * from "./bundles.sql";
+export * from "./daily-featured-projects.sql";
+export * from "./hall-of-fame.sql";
+export * from "./packages.sql";
+export * from "./project-trends.sql";
+export * from "./project-trends-view.sql";
+export * from "./projects.sql";
+export * from "./repo-trends.sql";
+export * from "./repo-trends-view.sql";
+export * from "./repos.sql";
+export * from "./snapshots.sql";
+export * from "./tags.sql";
